Initial commit
This is the initial commit. I'm not sure how I want to handle this in
the long term. Maybe have a normal and an _untested library. Once
symbols are tested in manufcturing, they could be moved to the normal
lib.

It would also be nice to have some automated way of running some
sanity checks on the symbols. For example, test that the pin start and
end are on a 100 mil grid.

These are some thoughts for the future. The way I will start this is
to include symbols for circuits that I build, and go from there.